Conflicts:
	ERP_WS/src/main/java/com/bap/erp/servicios/impl/rh/RhDescuentoEmpleadoCargoServiceImpl.java

+++ STARTUPWEEKEND/src/main/java/com/entrepreneur/startupweekend/modelo/GpsMessage.java
@@ -6,14 +6,14 @@ import java.math.BigDecimal;
 public class GpsMessage implements Serializable {
 
     private String serial;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
     public GpsMessage() {
         
     }
     
-    public GpsMessage(String serial, Long latitude, Long longitude) {
+    public GpsMessage(String serial, Double latitude, Double longitude) {
         this.serial = serial;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -27,23 +27,21 @@ public class GpsMessage implements Serializable {
         this.serial = serial;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 
     
-    
-
 }
